@@ -21,31 +21,34 @@ package io.jawg.osmcontributor.ui.dialogs;
 import android.accounts.AccountManager;
 import android.app.Activity;
 import android.app.DialogFragment;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.gms.common.AccountPicker;
 
 import org.greenrobot.eventbus.EventBus;
+import org.greenrobot.eventbus.Subscribe;
+import org.greenrobot.eventbus.ThreadMode;
 
 import javax.inject.Inject;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import butterknife.OnClick;
 import io.jawg.osmcontributor.OsmTemplateApplication;
 import io.jawg.osmcontributor.R;
+import io.jawg.osmcontributor.database.preferences.LoginPreferences;
+import io.jawg.osmcontributor.rest.events.GoogleAuthenticatedEvent;
 import io.jawg.osmcontributor.rest.security.GoogleOAuthManager;
-import io.jawg.osmcontributor.ui.events.login.AttemptLoginEvent;
-import io.jawg.osmcontributor.ui.events.login.LoginInitializedEvent;
+import io.jawg.osmcontributor.ui.events.login.UpdateGoogleCredentialsEvent;
+import io.jawg.osmcontributor.ui.managers.executor.GenericSubscriber;
+import io.jawg.osmcontributor.ui.managers.login.UpdateCredentialsIfValid;
 import io.jawg.osmcontributor.utils.StringUtils;
 
 /**
@@ -54,15 +57,6 @@ import io.jawg.osmcontributor.utils.StringUtils;
 public class LoginDialogFragment extends DialogFragment {
     private static final int PICK_ACCOUNT_CODE = 1;
 
-    @BindView(R.id.dialog_connection_google_button)
-    Button googleButton;
-
-    @BindView(R.id.dialog_connection_login_button)
-    Button loginButton;
-
-    @BindView(R.id.dialog_connection_login_later)
-    TextView loginLaterTextView;
-
     @BindView(R.id.dialog_connection_login_edit_text)
     EditText loginEditText;
 
@@ -70,13 +64,22 @@ public class LoginDialogFragment extends DialogFragment {
     EditText passwordEditText;
 
     @Inject
+    LoginPreferences loginPreferences;
+
+    @Inject
     GoogleOAuthManager googleOAuthManager;
 
-    private EventBus eventBus;
+    @Inject
+    EventBus eventBus;
 
-    public static LoginDialogFragment newInstance(EventBus bus) {
+    @Inject
+    UpdateCredentialsIfValid updateCredentialsIfValid;
+
+    private OnLoginSuccessfulListener onLoginSuccessfulListener;
+
+    public static LoginDialogFragment newInstance(OnLoginSuccessfulListener onLoginSuccessfulListener) {
         LoginDialogFragment fragment = new LoginDialogFragment();
-        fragment.setEventBus(bus);
+        fragment.onLoginSuccessfulListener = onLoginSuccessfulListener;
         return fragment;
     }
 
@@ -88,36 +91,20 @@ public class LoginDialogFragment extends DialogFragment {
 
         ((OsmTemplateApplication) getActivity().getApplication()).getOsmTemplateComponent().inject(this);
 
-        googleButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = AccountPicker.newChooseAccountIntent(
-                        null, null,
-                        new String[]{"com.google"},
-                        false, null, null, null, null);
-                startActivityForResult(intent, PICK_ACCOUNT_CODE);
-            }
-        });
+        eventBus.register(this);
 
+        loginEditText.setText(loginPreferences.retrieveLogin());
+        passwordEditText.setText(loginPreferences.retrievePassword());
 
-        loginLaterTextView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                dismiss();
-            }
-        });
-
-        loginButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (!StringUtils.isEmpty(loginEditText.getText()) && !StringUtils.isEmpty(passwordEditText.getText())) {
-                    eventBus.post(new AttemptLoginEvent(loginEditText.getText().toString(), passwordEditText.getText().toString()));
-                } else {
-                    Toast.makeText(getActivity(), R.string.empty_fields, Toast.LENGTH_LONG).show();
-                }
-            }
-        });
         return rootView;
+    }
+
+
+    @Override
+    public void onDestroy() {
+        eventBus.unregister(this);
+        updateCredentialsIfValid.unsubscribe();
+        super.onDestroy();
     }
 
     @Override
@@ -131,18 +118,64 @@ public class LoginDialogFragment extends DialogFragment {
         }
     }
 
-    @Override
-    public void onDismiss(DialogInterface dialog) {
-        super.onDismiss(dialog);
-        eventBus.postSticky(new LoginInitializedEvent());
+    @OnClick(R.id.dialog_connection_google_button)
+    void onGoogleButtonClicked() {
+        Intent intent = AccountPicker.newChooseAccountIntent(
+                null, null,
+                new String[]{"com.google"},
+                false, null, null, null, null);
+        startActivityForResult(intent, PICK_ACCOUNT_CODE);
     }
 
-    public void setEventBus(EventBus eventBus) {
-        this.eventBus = eventBus;
+    @OnClick(R.id.dialog_connection_login_button)
+    void onConnectionButtonClicked() {
+        if (!StringUtils.isEmpty(loginEditText.getText()) && !StringUtils.isEmpty(passwordEditText.getText())) {
+            updateCredentialsIfValid.init(loginEditText.getText().toString(), passwordEditText.getText().toString()).execute(new UpdateCredentialsIfValidObservable());
+        } else {
+            Toast.makeText(getActivity(), R.string.empty_fields, Toast.LENGTH_LONG).show();
+        }
     }
 
-    public void resetLoginFields() {
+    @OnClick(R.id.dialog_connection_login_later)
+    void onLoginLaterClicked() {
+        dismiss();
+    }
+
+    @Subscribe(threadMode = ThreadMode.MAIN)
+    public void onGoogleAuthenticatedEvent(GoogleAuthenticatedEvent event) {
+        if (event.isSuccessful()) {
+            eventBus.post(new UpdateGoogleCredentialsEvent(event.getToken(), event.getTokenSecret(), event.getConsumer(), event.getConsumerSecret()));
+        } else {
+            Toast.makeText(getActivity(), R.string.error_login, Toast.LENGTH_SHORT).show();
+        }
+        dismiss();
+    }
+
+    private void resetLoginFields() {
         loginEditText.setText("");
         passwordEditText.setText("");
+    }
+
+    public interface OnLoginSuccessfulListener {
+
+        void onLoginSuccessful();
+    }
+
+    private class UpdateCredentialsIfValidObservable extends GenericSubscriber<Boolean> {
+        @Override
+        public void onNext(Boolean isValidCredentials) {
+            if (isValidCredentials) {
+                Toast.makeText(getActivity(), R.string.valid_login, Toast.LENGTH_SHORT).show();
+
+                if (onLoginSuccessfulListener != null) {
+                    onLoginSuccessfulListener.onLoginSuccessful();
+                }
+
+                dismiss();
+            } else {
+                Toast.makeText(getActivity(), R.string.error_first_login, Toast.LENGTH_SHORT).show();
+                resetLoginFields();
+            }
+        }
     }
 }
