@@ -20,7 +20,9 @@ package io.jawg.osmcontributor.ui.managers;
 
 
 import android.app.Application;
+import android.content.SharedPreferences;
 
+import com.firebase.jobdispatcher.FirebaseJobDispatcher;
 import com.mapbox.mapboxsdk.geometry.LatLng;
 
 import org.greenrobot.eventbus.EventBus;
@@ -35,6 +37,7 @@ import java.util.Map;
 
 import javax.inject.Inject;
 
+import io.jawg.osmcontributor.R;
 import io.jawg.osmcontributor.database.dao.PoiNodeRefDao;
 import io.jawg.osmcontributor.model.entities.Action;
 import io.jawg.osmcontributor.model.entities.Condition;
@@ -52,21 +55,26 @@ import io.jawg.osmcontributor.ui.events.edition.PleaseApplyPoiPositionChange;
 import io.jawg.osmcontributor.ui.events.edition.PoiChangesApplyEvent;
 import io.jawg.osmcontributor.ui.events.map.PleaseCreateNoTagPoiEvent;
 import io.jawg.osmcontributor.ui.events.map.PoiNoTypeCreated;
+import io.jawg.osmcontributor.ui.managers.sync.PushToOSMService;
 import timber.log.Timber;
 
 public class EditPoiManager {
 
-    PoiManager poiManager;
-    Application application;
-    PoiNodeRefDao poiNodeRefDao;
-    EventBus eventBus;
+    private Application application;
+    private PoiManager poiManager;
+    private PoiNodeRefDao poiNodeRefDao;
+    private EventBus eventBus;
+    private FirebaseJobDispatcher dispatcher;
+    private SharedPreferences sharedPreferences;
 
     @Inject
-    public EditPoiManager(PoiManager poiManager, PoiNodeRefDao poiNodeRefDao, Application application, EventBus eventBus) {
-        this.poiManager = poiManager;
+    public EditPoiManager(Application application, PoiManager poiManager, PoiNodeRefDao poiNodeRefDao, EventBus eventBus, FirebaseJobDispatcher dispatcher, SharedPreferences sharedPreferences) {
         this.application = application;
+        this.poiManager = poiManager;
         this.poiNodeRefDao = poiNodeRefDao;
         this.eventBus = eventBus;
+        this.dispatcher = dispatcher;
+        this.sharedPreferences = sharedPreferences;
     }
 
     @Subscribe(threadMode = ThreadMode.ASYNC)
@@ -84,6 +92,7 @@ public class EditPoiManager {
             editPoi.setUpdated(true);
             poiManager.savePoi(editPoi);
             poiManager.updatePoiTypeLastUse(editPoi.getType().getId());
+            schedulePushJob();
         }
 
         eventBus.post(new PoiChangesApplyEvent());
@@ -101,6 +110,7 @@ public class EditPoiManager {
         editPoi.setUpdated(true);
         poiManager.savePoi(editPoi);
         poiManager.updatePoiTypeLastUse(editPoi.getType().getId());
+        schedulePushJob();
     }
 
     @Subscribe(threadMode = ThreadMode.ASYNC)
@@ -117,6 +127,7 @@ public class EditPoiManager {
         poiNodeRef.setLatitude(newLatLng.getLatitude());
         poiNodeRef.setUpdated(true);
         poiNodeRefDao.createOrUpdate(poiNodeRef);
+        schedulePushJob();
     }
 
     @Subscribe(threadMode = ThreadMode.ASYNC)
@@ -129,6 +140,7 @@ public class EditPoiManager {
         poiManager.savePoi(poi);
         poiManager.updatePoiTypeLastUse(poi.getType().getId());
         eventBus.post(new PoiChangesApplyEvent());
+        schedulePushJob();
     }
 
     @Subscribe(threadMode = ThreadMode.ASYNC)
@@ -146,6 +158,7 @@ public class EditPoiManager {
             poi.setToDelete(true);
             poiManager.savePoi(poi);
         }
+        schedulePushJob();
     }
 
     @Subscribe(threadMode = ThreadMode.ASYNC)
@@ -174,6 +187,13 @@ public class EditPoiManager {
         poiManager.updatePoiTypeLastUse(event.getPoiType().getId());
 
         eventBus.post(new PoiNoTypeCreated());
+        schedulePushJob();
+    }
+
+    private void schedulePushJob() {
+        if (sharedPreferences.getBoolean(application.getString(R.string.shared_prefs_auto_commit), false)) {
+            PushToOSMService.schedulePushJob(dispatcher);
+        }
     }
 
     private Long saveOldVersionOfPoi(Poi poi) {
